+++ qase-javascript-commons/changelog.md
@@ -1,3 +1,10 @@
+# qase-javascript-commons@2.2.17
+
+## What's new
+
+- Logging of host system details to improve debugging and traceability.
+- Output of installed packages in logs for better environment visibility.
+
 # qase-javascript-commons@2.2.16
 
 ## What's new